+++ doc/changelog.md
@@ -1,6 +1,8 @@
 ## 0.9.6
 #### element.js Changes
  - Added some input filtering to prevent unneeded warnings from the parse function
+ - **Bugfix:** in the parse that caused malformed strings split into a list instead of failing due to being malformed
+ - **Bugfix:** str's with comma's were getting malformed when in dict or array
 #### Fallen Changes
  - Added new `img` element in `staging.js`